Remove proxy class
