+++ lib/hackney/income/tenancy_classification/helpers/agreement_helpers.rb
@@ -5,22 +5,22 @@ module Hackney
         module AgreementHelpers
           include HelpersBase
 
-          def active_agreement_ma?
+          def active_agreement?
             most_recent_agreement.present? && most_recent_agreement.active?
           end
 
-          def informal_breached_agreement_ma?
+          def informal_breached_agreement?
             breached_agreement? && !court_breach_agreement?
           end
 
-          def breached_agreement_ma?
+          def breached_agreement?
             return false if most_recent_agreement.blank?
             return false if most_recent_agreement.start_date.blank?
 
             most_recent_agreement.breached?
           end
 
-          def court_breach_agreement_ma?
+          def court_breach_agreement?
             return false unless breached_agreement?
             return false unless most_recent_agreement.formal?
             return false if most_recent_court_case.court_date.blank?
